Split options into base command class
