+++ core/lib/Drupal/Core/Command/DbDumpCommand.php
@@ -9,10 +9,7 @@ namespace Drupal\Core\Command;
 
 use Drupal\Component\Utility\Variable;
 use Drupal\Core\Database\Connection;
-use Drupal\Core\Database\Database;
-use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
@@ -31,7 +28,7 @@ use Symfony\Component\Console\Output\OutputInterface;
  *
  * @see \Drupal\Core\Command\DbDumpApplication
  */
-class DbDumpCommand extends Command {
+class DbDumpCommand extends DbCommandBase {
 
   /**
    * An array of table patterns to exclude completely.
@@ -54,48 +51,8 @@ class DbDumpCommand extends Command {
    */
   protected function configure() {
     $this->setName('dump-database-d8-mysql')
-      ->setDescription('Dump the current database to a generation script')
-      ->addOption('database', NULL, InputOption::VALUE_OPTIONAL, 'The database connection name to use.', 'default')
-      ->addOption('database-url', 'db-url', InputOption::VALUE_OPTIONAL, 'A database url to parse and use as the database connection.')
-      ->addOption('prefix', NULL, InputOption::VALUE_OPTIONAL, 'Override or set the table prefix used in the database connection.');
-  }
-
-  /**
-   * Parse input options decide on a database.
-   *
-   * @todo this could probably be refactored to use global connections.
-   *
-   * @param \Symfony\Component\Console\Input\InputInterface $input
-   *   Input object.
-   * @return \Drupal\Core\Database\Connection
-   */
-  protected function getDatabaseConnection(InputInterface $input) {
-
-    // Load connection from a url.
-    if ($input->getOption('database-url')) {
-      // ensure database connection isn't set.
-      if (Database::getConnectionInfo('db-tools')) {
-        throw new \RuntimeException('Database "db-tools" is already defined. Can not define database provided.');
-      }
-      $info = Database::parseConnectionInfo($input->getOption('database-url'));
-      Database::addConnectionInfo('db-tools', 'default', $info);
-      $key = 'db-tools';
-    }
-    else {
-      $key = $input->getOption('database');
-    }
-
-    // If they supplied a prefix, replace it in the connection information.
-    $prefix = $input->getOption('prefix');
-    if ($prefix) {
-      $info = Database::getConnectionInfo($key)['default'];
-      $info['prefix']['default'] = $prefix;
-
-      Database::removeConnection($key);
-      Database::addConnectionInfo($key, 'default', $info);
-    }
-
-    return Database::getConnection($key);
+      ->setDescription('Dump the current database to a generation script');
+    parent::configure();
   }
 
   /**
